the delete works!

+++ rest_server.mjs
@@ -6,7 +6,7 @@ import { default as sqlite3 } from 'sqlite3';
 
 const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
 console.log(__dirname)
-const db_filename = path.join(__dirname, 'db', 'stpaul_crime.sqlite3');
+const db_filename = path.join(__dirname, 'db', 'stpaul_crime - Copy.sqlite3');
 console.log(db_filename)
 
 const port = 8000;
@@ -310,39 +310,39 @@ app.put('/new-incident', (req, res) => {
 });
 
 // DELETE request handler for new crime incident
-// app.delete('/remove-incident', (req, res) => {
-//     dbSelect('SELECT * FROM Incidents WHERE case_number = ?', [req.body])
-//         .then((rows) => {
-//             if (rows.length === 0) {
-//                 res.status(500).type('txt').send('Case number not found');
-//             } else {
-//                 const sql = 'DELETE FROM Incidents WHERE case_number = ?';
-//                 dbRun(sql, [case_number])
-//                     .then(() => {
-//                         res.status(200).type('txt').send('OK');
-//                     })
-//                     .catch((error) => {
-//                         res.status(500).type('txt').send(error);
-//                     });
-//             }
-//         })
-//         .catch((error) => {
-//             res.status(500).type('txt').send(error);
-//         });
-// });
-
-
 app.delete('/remove-incident', (req, res) => {
-    console.log(req.body.case_number);
-    const sql = 'DELETE FROM Incidents WHERE case_number = ?';
-    dbRun(sql, req.body.case_number)
-        .then(() => {
-            res.status(200).type('txt').send('OK');
+    dbSelect('SELECT * FROM Incidents WHERE case_number = ?', [req.body.case_number])
+        .then((rows) => {
+            if (rows.length === 0) {
+                res.status(500).type('txt').send('Case number not found');
+            } else {
+                const sql = 'DELETE FROM Incidents WHERE case_number = ?';
+                dbRun(sql, [req.body.case_number])
+                    .then(() => {
+                        res.status(200).type('txt').send('OK');
+                    })
+                    .catch((error) => {
+                        res.status(500).type('txt').send(error);
+                    });
+            }
         })
         .catch((error) => {
             res.status(500).type('txt').send(error);
         });
 });
+
+
+// app.delete('/remove-incident', (req, res) => {
+//     console.log(req.body.case_number);
+//     const sql = 'DELETE FROM Incidents WHERE case_number = ?';
+//     dbRun(sql, req.body.case_number)
+//         .then(() => {
+//             res.status(200).type('txt').send('OK');
+//         })
+//         .catch((error) => {
+//             res.status(500).type('txt').send(error);
+//         });
+// });
